refactor(utils): extract error classes to dedicated module
