+++ packages/utils/src/prettyBytes.ts
@@ -4,7 +4,7 @@
  * MIT Licensed
  */
 
-import { ValidationError } from "./types.ts";
+import { ValidationError } from "./error.ts";
 
 const UNITS = ["B", "kB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"] as const;
 type Unit = (typeof UNITS)[number];
